Added menu button for objects dropping
