+++ src/js/shot-generator/SelectionManager.js
@@ -16,8 +16,6 @@ const {
   getActiveCamera
 } = require('../shared/reducers/shot-generator')
 
-const { dropObject, dropCharacter } = require("../utils/dropToObjects")
-
 function getObjectsFromIcons ( objects ) {
   return objects
       // visible objects
@@ -151,11 +149,6 @@ const SelectionManager = connect(
     o.userData.type === 'boneControl' ||
     (useIcons && o.isPerspectiveCamera)
   )
-
-  const dropingPlaces = scene.children.filter(o =>
-    o.userData.type === 'object' ||
-    o.userData.type === 'character' ||
-    o.userData.type === 'ground')
 
   const mouse = event => {
     const rect = el.getBoundingClientRect()
@@ -533,41 +526,16 @@ const SelectionManager = connect(
     setLastDownId(null)
   }
 
-  const onDropKeyPressed = event => {
-    if( event.ctrlKey ) {
-      if( event.key === 'b' ) {
-        let changes = {}
-        for( let i = 0; i < selections.length; i++ ) {
-          let selection = scene.children.find( child => child.userData.id === selections[i] )
-          if( selection.userData.type === "object" ) {
-            dropObject( selection, dropingPlaces )
-            let pos = selection.position
-            changes[ selections[i] ] = { x: pos.x, y: pos.z, z: pos.y }
-          } else if ( selection.userData.type === "character" ) {
-            dropCharacter( selection, dropingPlaces )
-            let pos = selection.position
-            changes[ selections[i] ] = { x: pos.x, y: pos.z, z: pos.y }
-          }
-        }
-        updateObjects(changes)
-      }
-    }
-  }
-
   useLayoutEffect(() => {
     el.addEventListener('pointerdown', onPointerDown)
     el.addEventListener('pointermove', onPointerMove)
     document.addEventListener('pointerup', onPointerUp)
-    if( !useIcons )
-      window.addEventListener('keydown', onDropKeyPressed)
     return function cleanup () {
       el.removeEventListener('pointerdown', onPointerDown)
       el.removeEventListener('pointermove', onPointerMove)
       document.removeEventListener('pointerup', onPointerUp)
-      if( !useIcons )
-        window.removeEventListener('keydown', onDropKeyPressed)
     }
-  }, [onPointerDown, onPointerUp, onPointerMove, onDropKeyPressed])
+  }, [onPointerDown, onPointerUp, onPointerMove])
 
   useLayoutEffect(() => {
     if (dragTarget) {
